Add mention of recursive import support to readme
First off, THANK YOU for creating this library.  It's exactly what I've been looking for!

I just wanted to mention that you may want to mention that rigger does indeed support recursive imports in your readme, as that was a required feature for the current task I'm working on, and I had to go looking through your closed issues & commits to verify that it was indeed supported.  It's a small change, but could make the difference between someone adopting this library or not.